+++ vda5050_core/CMakeLists.txt
@@ -106,6 +106,7 @@ vda5050_configure_target_alias(
 target_link_libraries(vda5050_validation PUBLIC
   ${PROJECT_NAME}::types
   ${PROJECT_NAME}::errors
+  ${PROJECT_NAME}::layout
   fmt::fmt
 )
 target_include_directories(vda5050_validation
@@ -345,20 +346,16 @@ if(BUILD_TESTING)
 
   ament_add_gmock(test_validation
     test/validation/test_unit_content_validator.cpp
-    test/validation/test_unit_pre_send_validator.cpp
     test/validation/test_unit_order_graph_validator.cpp
+    test/validation/test_unit_pre_send_validator.cpp
+    test/validation/test_unit_traversability_validator.cpp
+    test/validation/test_unit_capability_validator.cpp
+    test/validation/test_unit_protocol_limits_validator.cpp
+    test/validation/test_unit_action_conflict_validator.cpp
+    test/validation/test_unit_instant_action_mode_validator.cpp
+    test/validation/test_unit_factsheet_alignment.cpp
   )
   target_link_libraries(test_validation vda5050_core::validation)
-
-  ament_add_gmock(test_master_validation_order_action
-    test/master/test_unit_traversability_validator.cpp
-    test/master/test_unit_capability_validator.cpp
-    test/master/test_unit_protocol_limits_validator.cpp
-    test/master/test_unit_action_conflict_validator.cpp
-    test/master/test_unit_instant_action_mode_validator.cpp
-    test/master/test_unit_factsheet_alignment.cpp
-  )
-  target_link_libraries(test_master_validation_order_action vda5050_core::master)
 
   ament_add_gmock(test_layout_loader
     test/layout/test_unit_loader.cpp
