+++ vda5050_core/include/vda5050_core/validation/instant_action_mode_validator.hpp
@@ -16,20 +16,20 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__MASTER__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
-#define VDA5050_CORE__MASTER__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
+#ifndef VDA5050_CORE__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
+#define VDA5050_CORE__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
 
 #include <string>
 
-#include "vda5050_core/master/validation/pre_send_validator.hpp"
-#include "vda5050_core/order_utils/validation_result.hpp"
+#include "vda5050_core/validation/pre_send_validator.hpp"
+#include "vda5050_core/errors/validation_result.hpp"
 #include "vda5050_core/types/instant_actions.hpp"
 
-namespace vda5050_core::master {
+namespace vda5050_core::validation {
 
 /// \brief Operating-mode gate: in non-AUTOMATIC modes, only the predefined
 /// instant-scope actions (is_mode_exempt_action_type) may be sent.
-vda5050_core::order_utils::ValidationResult validate_instant_action_mode(
+vda5050_core::errors::ValidationResult validate_instant_action_mode(
   const PreSendContext& ctx,
   const vda5050_core::types::InstantActions& actions);
 
@@ -37,6 +37,6 @@ vda5050_core::order_utils::ValidationResult validate_instant_action_mode(
 /// the operating-mode gate (e.g. stateRequest, cancelOrder, initPosition).
 bool is_mode_exempt_action_type(const std::string& action_type);
 
-}  // namespace vda5050_core::master
+}  // namespace vda5050_core::validation
 
-#endif  // VDA5050_CORE__MASTER__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
+#endif  // VDA5050_CORE__VALIDATION__INSTANT_ACTION_MODE_VALIDATOR_HPP_
