+++ vda5050_core/src/validation/capability_validator.cpp
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-#include "vda5050_core/master/validation/capability_validator.hpp"
+#include "vda5050_core/validation/capability_validator.hpp"
 
 #include <algorithm>
 #include <functional>
@@ -26,13 +26,13 @@
 #include "vda5050_core/errors/error_codes.hpp"
 #include "vda5050_core/errors/error_factory.hpp"
 
-namespace vda5050_core::master {
+namespace vda5050_core::validation {
 
 namespace {
 
 using ::vda5050_core::errors::CapabilityValidationError;
 using ::vda5050_core::errors::create_error;
-using ::vda5050_core::order_utils::ValidationResult;
+using ::vda5050_core::errors::ValidationResult;
 using ::vda5050_core::types::AGVAction;
 using ::vda5050_core::types::ErrorReference;
 
@@ -206,4 +206,4 @@ ValidationResult validate_capability(
   return res;
 }
 
-}  // namespace vda5050_core::master
+}  // namespace vda5050_core::validation
