+++ vda5050_core/test/validation/test_unit_protocol_limits_validator.cpp
@@ -22,12 +22,12 @@
 #include <string>
 
 #include "vda5050_core/errors/error_codes.hpp"
-#include "vda5050_core/master/validation/protocol_limits_validator.hpp"
+#include "vda5050_core/validation/protocol_limits_validator.hpp"
 
 namespace {
 
 ::testing::AssertionResult AllErrorsHaveProtocolLimitType(
-  const vda5050_core::order_utils::ValidationResult& res)
+  const vda5050_core::errors::ValidationResult& res)
 {
   for (const auto& e : res.errors)
   {
@@ -42,7 +42,7 @@ namespace {
 }
 
 ::testing::AssertionResult AnyErrorMentions(
-  const vda5050_core::order_utils::ValidationResult& res,
+  const vda5050_core::errors::ValidationResult& res,
   const std::string& needle)
 {
   for (const auto& e : res.errors)
@@ -60,7 +60,9 @@ namespace {
 
 }  // namespace
 
-namespace vda5050_core::master::test {
+namespace vda5050_core::validation::test {
+
+using vda5050_core::master::AGVState;
 
 namespace {
 
@@ -235,4 +237,4 @@ TEST(ProtocolLimitsValidatorTest, MultipleLimitErrorsAccumulate)
   EXPECT_TRUE(AnyErrorMentions(res, "node_actions"));
 }
 
-}  // namespace vda5050_core::master::test
+}  // namespace vda5050_core::validation::test
